requirements.txt
aioconsole==0.5.1
altair==4.2.0
asyncio==3.4.3
backports.zoneinfo==0.2.1
blinker==1.5
cachetools==5.2.0
click==8.1.3
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
fonttools==4.38.0
gitdb==4.0.9
GitPython==3.1.27
importlib-metadata==4.12.0
importlib-resources==5.9.0
Jinja2==3.1.2
jsonschema==4.16.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
numpy==1.24.1
onnxruntime-tidl @ file:///host/home/edgeai-user/a0497747local/test-build/rootfs/home/root/arago_j7_pywhl/onnxruntime_tidl-1.7.0-cp38-cp38-linux_aarch64.whl
opencv-python==4.7.0.68
packaging==23.0
pandas==1.4.4
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.2
pyarrow==9.0.0
pydeck==0.8.0b3
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pypylon==1.8.0
pyrsistent==0.18.1
pyserial==3.5
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
rich==12.5.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.12.2
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
validators==0.20.0
watchdog==2.1.9
websockets==10.4
zipp==3.8.1
